Add files via upload
Enhanced EfficientNetB6 model with extra layers for better performance. We added a global average pooling layer to reduce the number of parameters 
and simplify the feature maps, making the model more efficient and less likely to overfit. This layer connects the main model to the newly added layers. 
We also used batch normalisation to keep the activations balanced and improve training stability. This procedure accelerates convergence during training, improves model stability, and lessens internal covariate movements. The first mini-batch mean, and variance were determined using the formula below.
μ=(1/m)×∑X						 
σ^2=(1/m)mean,		     				  
Here, μ,σ^2,and m are the mean, variance, and the number of instances in mini-batch respectively. 
X_h=(X-μ) √(σ^2+ε)					
Here ε represents the numerical stability of a small constant. Additionally, the use of shift and scale empowers the neural network to adopt the normalised activation for certain tasks. This is accomplished by adding the learnable parameters betta and gamma, which are displayed in equation 
Y=γ×X+β						  
This method uses backpropagation to update β and γ as learnable parameters during training. Each layer's scaling and shifting are optimized. In addition, three dense layers were incorporated, each with varying numbers of units, ReLU activation, and dropout, as outlined in the configuration as shown in Figure 4. The initial dense layer consists of 128 units, employs ReLU activation, and includes a dropout rate of 0.3. The subsequent dense layer contains 64 units, also uses ReLU activation, and has a dropout rate of 0.2.The third dense layer has 32 units with ReLU activation and a dropout rate of 0.2. The ReLU function. To assist the network in acquiring more generic and dependable characteristics, the dropout layer randomly turns off certain neurones during training, preventing overfitting.